ProcessController and related code refactor (#948)
This notably aim at fixing the following issues :
- #642
- #940
- #941 
- #942 

Main change is reverting this commit :
ac24298f56a7111e76d561c673e916dc840ee433 (note that other related
changes were made following this commit as it broke a bunch of stuff)

Originally, ProcessControllers enabled state toggling (either by the
user or by a Reliability module) was done by toggling the
pseudo-resource `flowState`. The mentioned commit changed that logic to
instead alter the pseudo-resource amount (setting it either zero or
maxAmount), which notably broke self-consuming processes (see issue
#940), and more generally made things a lot messier.

Aside from fixing issue #940, reverting this change was necessary to
implement the fix for issue #642, and more generally it make the overall
implementation a lot cleaner.

The PR implement a save/craft persisted data migration code to preserve
backward compatibility with existing saves from 3.24 and below. However,
do note that once migrated to 3.25 or above, the save will not work
correctly anymore in 3.24 and below, so backing up saves is recommended
if for any reason people want to downgrade. This being said, the issues
should be relatively minor, mainly user-disabled or reliability-broken
PCs being forever stuck in the disabled state.

The PR implement a proper fix for issue #642 by allowing the
`IConfigurable.Configure()` method to specify a multiplier for when
multiple slots are assigned to the same setup. The ProcessController
module then implement proper handling of the multiplier by applying it
to the pseudo-resource amount/maxAmount in its
`IConfigurable.Configure()` implementation.

Due to the changes being somehwat risky, this should probably be
playtested a bit before merging/releasing.

Dev builds (remember to backup your save as it won't be compatible with
3.24 anymore) :

[Kerbalism-Core_dev-9386_for_KSP1.8.0-1.12.9.zip](https://github.com/user-attachments/files/22296982/Kerbalism-Core_dev-9386_for_KSP1.8.0-1.12.9.zip)

[Kerbalism-Config_dev-9386_for_KSP1.8.0-1.12.9.zip](https://github.com/user-attachments/files/22296957/Kerbalism-Config_dev-9386_for_KSP1.8.0-1.12.9.zip)